fix: move retry policies to networking
