+++ crates/rattler-bin/src/commands/create.rs
@@ -11,7 +11,9 @@ use rattler_conda_types::{
     Channel, ChannelConfig, GenericVirtualPackage, MatchSpec, PackageRecord, Platform,
     PrefixRecord, RepoDataRecord, Version,
 };
-use rattler_networking::{default_retry_policy, AuthenticatedClient, AuthenticationStorage};
+use rattler_networking::{
+    retry_policies::default_retry_policy, AuthenticatedClient, AuthenticationStorage,
+};
 use rattler_repodata_gateway::fetch::{
     CacheResult, DownloadProgress, FetchRepoDataError, FetchRepoDataOptions,
 };
